test: move applyDisableRules specs to options-list.spec; scope about-page e2e selectors
The applyDisableRules suites test a pure function in options-list.ts, so
they belong in its spec (AGENTS.md: specs mirror sources). The about-page
e2e now targets #version / #github a / #banner span with an anchored
version regex and an exact href instead of text= and substring matches.
